Refactor for sanity
