+++ test/parser_test.rb
@@ -1,8 +1,5 @@
-require "minitest/autorun"
-require "minitest/pride"
-require "pry"
-require_relative "test_helper"
-require_relative "../lib/parser"
+require "./test/test_helper"
+require "./lib/parser"
 
 class ParserTest < Minitest::Test
