+++ test/path_respond_test.rb
@@ -1,7 +1,5 @@
-require "minitest/autorun"
-require "minitest/pride"
-require "pry"
-require_relative "../lib/path_respond"
+require "./test/test_helper"
+require "./lib/path_respond"
 
 class PathRespondTest < Minitest::Test
